requirements.txt
# requirements.txt
google-genai
google-generativeai
google-cloud-aiplatform
requests
python-dotenv
mcp
fastapi
uvicorn
pytest
discord.py